requirements.txt
web3>=6.11,<7
eth-account>=0.10.0,<0.11
requests>=2.31
python-dotenv>=1.0
rich>=13.7
websocket-client>=1.7
InquirerPy>=0.3.4
pyfiglet>=0.8.post1
colorama>=0.4.6
